Add new game and battle screen constants; refactor button initialization in Game class
