+++ constants.py
@@ -38,6 +38,18 @@ class Game_Constants:
     POPUP_WIDTH = 400
     POPUP_HEIGHT = 200
 
+    FPS = 60
+
+    # New Game Constants
+    NEW_GAME_SPACING = SCREEN_HEIGHT // 30
+
+    NEW_GAME_BUTTON_WIDTH = SCREEN_WIDTH // 8 * 3
+    NEW_GAME_BUTTON_HEIGHT = SCREEN_HEIGHT // 12
+    NEW_GAME_BUTTON_X = SCREEN_WIDTH // 2 - NEW_GAME_BUTTON_WIDTH // 2
+    NEW_GAME_BUTTON_Y = SCREEN_HEIGHT // 2 - NEW_GAME_BUTTON_HEIGHT // 2
+    NEW_GAME_BUTTON_INCREMENT = NEW_GAME_BUTTON_HEIGHT + NEW_GAME_SPACING
+
+    # Battle Screen Constants
     BATTLE_SCREEN_SPACING = SCREEN_HEIGHT // 30
 
     BATTLE_SCREEN_BUTTON_BORDER_X = 0
